acmens.py: fallback to `dns`
Sometimes `http` isn't available as a challenge type. E.g. for wildcards
or when `dns` was used for the domain lately.

Instead of being forced to use `--challenge dns` for wildcard
certificates, the `--challenge` flag can be omitted now and it will
automatically fallback to `dns`.

This is especially useful for SAN certificates, where it can fallback to
the other challenge type, if the preferred type isn't available for one
of the multiple domains.